Original ackmode patch.
This seems to be pretty close to the original patched commit. I only have
access to the zip file that contains the modified sources. To actually
track down the true commit I'd need more information or a copy of the
entire repository.

As it stands, this seems good enough to understand what the original patch
did well enough to port to DW 1.7.

My initial take is that it's pretty straightforward. When a KISS frame
is coming in, the 'cmd' byte is checked to see if it's of the special
XKISS_CMD_DATA type, and if so it's treated like a regular data frame
with one exception - the first two bytes (uint16_t) are some kind of token
or sequence number that is stripped from the packet. The packet struct
is expanded to include the original 'cmd' byte and that ACK token/value,
as well as the client socket/connection that the packet came in on.

When a packet is transmitted, the saved 'cmd' byte is checked to see if
the packet was original an ack mode packet, if so then a small KISS 'reply'
is constructed and buffered that includes the same ack mode cmd value and
then the 2 byte ack token/value. Because multiple packets can be
transmitted in the loop this is done from, multiple replies are buffered
and sent back at once. They get the usual KISS framing.

It appears that a special 'low level' or raw kiss send had to be included
to support the final sending of that reply.